+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Configurable frequency parameter to adjust log-server rate limiting
+- Add `frequency` parameter to adjust log-server rate limiting
+- Add `read_from_beginning` parameter to select log-server initial behavior
 
 ### Changed